Addressed Adelynn's Comments
