+++ All In/Features/Home/Views/FrequentAskedQuestion.swift
@@ -7,71 +7,15 @@
 
 import SwiftUI
 
-struct FAQ: Identifiable {
-    let id = UUID()
-    let question: String
-    let answer: String
-}
-
 struct FrequentAskedQuestion: View {
-
-    private let faqData = [
-        FAQ(
-            question: "What is sports betting?",
-            answer: """
-            Sports betting is the act of placing a wager on the outcome of a sporting event. \
-            It allows individuals to bet on various aspects of a game, such as which team \
-            will win, the final score, or specific player performances.
-            """
-        ),
-        FAQ(
-            question: "How do we place a bet?",
-            answer: """
-            Bets are placed in Home or Marketplace. In Home, you have the ability to buy \
-            "packs" by rarity type or player. There are four different levels of rarity \
-            which are common, rare, epic, and legendary. You can choose to buy a pack \
-            under a specific player of your choosing.
-
-            In marketplace, you can buy an event that is being sold secondhand from another \
-            user who opened up a rarity or player pack that you want to pursue.
-
-            When you choose a pack in Home or a contract in Marketplace, you will see they \
-            in your cart organized by the screen you bought them from. Once you place your \
-            bet, you will see all of your bets that are pending on the results in your Bet \
-            Tracker under Active Bets.
-            """
-        ),
-        FAQ(
-            question: "Do you keep your profit?",
-            answer: """
-            All costs and gains you make are fictional and do not correlate with \
-            non-fictional currency.
-            """
-        ),
-        FAQ(
-            question: "What happens if a game is canceled?",
-            answer: """
-            All games that are postponed or canceled will result in a refund in the \
-            original amount you spent to obtain that pack or contract.
-            """
-        ),
-        FAQ(
-            question: "How do I report issues?",
-            answer: """
-            Please email us at cornellappdev@gmail.com if you are experiencing any \
-            issues or bugs.
-            """
-        )
-    ]
-
     @State private var expandedQuestionIds = Set<UUID>()
 
     var myGradient = Gradient(
         colors: [
-          Constants.Colors.gradientBlue,
-          Constants.Colors.gradientLightBlue,
-          Constants.Colors.gradientPurple,
-          Constants.Colors.gradientLavender
+            Constants.Colors.gradientBlue,
+            Constants.Colors.gradientLightBlue,
+            Constants.Colors.gradientPurple,
+            Constants.Colors.gradientLavender
         ]
     )
 
@@ -80,9 +24,9 @@ struct FrequentAskedQuestion: View {
             ScrollView {
                 VStack(alignment: .leading, spacing: 20) {
                     header
-                    
+
                     questionsList
-                    
+
                     Spacer(minLength: 20)
                 }
                 .padding(.horizontal)
@@ -93,7 +37,10 @@ struct FrequentAskedQuestion: View {
             .scrollIndicators(.hidden)
         }
     }
+}
 
+// MARK: - Components
+extension FrequentAskedQuestion {
     private var header: some View {
         VStack(alignment: .leading, spacing: 8) {
             Text("How can we help?")
@@ -109,12 +56,15 @@ struct FrequentAskedQuestion: View {
 
     private var questionsList: some View {
         VStack(spacing: 16) {
-            ForEach(faqData) { faq in
+            ForEach(FAQ.sampleData) { faq in
                 questionItem(faq)
             }
         }
     }
+}
 
+// MARK: - Functions
+extension FrequentAskedQuestion {
     private func questionItem(_ faq: FAQ) -> some View {
         let isExpanded = expandedQuestionIds.contains(faq.id)
 
@@ -159,7 +109,14 @@ struct FrequentAskedQuestion: View {
         .clipShape(RoundedRectangle(cornerRadius: 16))
         .overlay(
             RoundedRectangle(cornerRadius: 16)
-                .stroke(LinearGradient(gradient: myGradient, startPoint: .topLeading, endPoint: .bottomTrailing), lineWidth: 1)
+                .stroke(
+                    LinearGradient(
+                        gradient: myGradient,
+                        startPoint: .topLeading,
+                        endPoint: .bottomTrailing
+                    ),
+                    lineWidth: 1
+                )
         )
     }
 }
